+++ gwcelery/tests/test_tasks_p_astro_gstlal.py
@@ -1,7 +1,6 @@
 import json
 from unittest.mock import patch
 
-import numpy as np
 import pkg_resources
 import pytest
 
@@ -46,34 +45,6 @@ def coinc_bytes_3():
         __name__, 'data/coinc.xml')
 
 
-@pytest.fixture(autouse=True, scope='module')
-def mock_ranking_pdf():
-    old = app.conf['p_astro_gstlal_ranking_pdf']
-    app.conf['p_astro_gstlal_ranking_pdf'] = pkg_resources.resource_filename(
-        __name__, 'data/ranking_data_G322589.xml.gz')
-    yield
-    app.conf['p_astro_gstlal_ranking_pdf'] = old
-
-
-def test_get_ln_f_over_b(ranking_data_bytes):
-    """Test to check `_get_ln_f_over_b` returns values
-    which are not inf or nan for high ln_likelihood_ratio"""
-    ln_f_over_b = p_astro_gstlal._get_ln_f_over_b(
-        ranking_data_bytes, [100., 200., 300.])
-    assert np.all(np.isfinite(ln_f_over_b))
-
-
-def test_get_event_ln_likelihood_ratio_svd_endtime_mass(coinc_bytes_1):
-    likelihood, mass1, mass2, spin1z, spin2z, snr, far = \
-        p_astro_gstlal._get_event_ln_likelihood_ratio_svd_endtime_mass(
-            coinc_bytes_1)
-    assert mass1 == pytest.approx(2.8, abs=0.1)
-    assert mass2 == pytest.approx(1.0, abs=0.1)
-    assert spin1z == pytest.approx(-0.99, abs=0.01)
-    assert spin2z == pytest.approx(0.049, abs=0.01)
-    assert likelihood == pytest.approx(21.65, abs=0.1)
-
-
 def test_compute_p_astro_1(coinc_bytes_1, ranking_data_bytes, mock_url):
     """Test to call `compute_p_astro` on gracedb event G322589.
     m1 = 2.7, m2 = 1.0 solar mass for this event"""
@@ -96,7 +67,7 @@ def test_compute_p_astro_2(coinc_bytes_2, ranking_data_bytes, mock_url):
 
     p_astros = json.loads(p_astro_gstlal.compute_p_astro(files))
 
-    assert p_astros['Terrestrial'] > 0.30
+    assert p_astros['Terrestrial'] > 0.25
 
 
 def test_failing_compute_p_astro(coinc_bytes_3, ranking_data_bytes, mock_url):